Cross-platform support for message sends with the GNUstep runtime.
Add feature gates for the platform specific objc_msgSend implementation
used by the GNUstep runtime (mips not currently supported because I'm
unaware of the calling conventions). In all other cases, fall back to the
cross-platform two-stage message sending.

This now works and passes some of the tests (I had a buggy old gcc version
yesterday, it seems). Every test that assumes NSObject to be present fails,
though because that is not part of the GNUstep runtime. We'll either have
to link gnustep-base for the tests to run properly, or provide stub
implementation.